modified alert dialog
